Crate zenobuf_core

Zenobuf Core - A simpler ROS-like framework in Rust

This crate provides the core functionality for the Zenobuf framework, a simpler ROS-like framework in Rust that uses Zenoh for transport and Protocol Buffers for serialization.

Re-exports

pub use client::Client;

pub use error::Error;

pub use error::Result;

pub use message::Message;

pub use node::Node;

pub use node::PublisherHandle;

pub use node::SubscriberHandle;

pub use node::ServiceHandle;

pub use node::ClientHandle;

pub use node::DropGuard;

pub use parameter::Parameter;

pub use publisher::Publisher;

pub use qos::QosProfile;

pub use qos::QosPreset;

pub use service::Service;

pub use subscriber::Subscriber;

pub use transport::ZenohTransport;

pub use transport::Transport;

Modules

client

Client implementation for Zenobuf

error

Error types for the Zenobuf framework

message

Message trait and utilities for working with Protocol Buffer messages

node

Node abstraction for Zenobuf

parameter

Parameter system for Zenobuf

publisher

Publisher implementation for Zenobuf

qos

Quality of Service (QoS) profiles for Zenobuf

service

Service implementation for Zenobuf

subscriber

Subscriber implementation for Zenobuf

time

Time utilities for Zenobuf

transport

Transport layer abstraction for Zenobuf