zenobuf_core::error

Enum Error

pub enum Error {

    Transport {
        source: Error,
        context: String,
    },
    MessageSerialization {
        source: EncodeError,
        type_name: &'static str,
    },
    MessageDeserialization {
        source: DecodeError,
        type_name: &'static str,
    },
    Serialization(String),
    Encoding(EncodeError),
    Decoding(DecodeError),
    NodeAlreadyExists {
        name: String,
    },
    TopicAlreadyExists {
        topic: String,
        node: String,
    },
    ServiceAlreadyExists {
        service: String,
        node: String,
    },
    ServiceCallTimeout {
        service: String,
        timeout_ms: u64,
    },
    ServiceCallFailed {
        service: String,
        reason: String,
    },
    Parameter {
        name: String,
        reason: String,
    },
    Node {
        node: String,
        reason: String,
    },
    Publisher {
        topic: String,
        reason: String,
    },
    Subscriber {
        topic: String,
        reason: String,
    },
    Service {
        service: String,
        reason: String,
    },
    Client {
        service: String,
        reason: String,
    },
    NotSupported {
        operation: String,
        reason: String,
    },
    NotImplemented {
        operation: String,
        reason: String,
    },
    Configuration {
        reason: String,
    },
    Network {
        reason: String,
    },
    Other {
        reason: String,
    },
    NodeAlreadyExistsLegacy(String),
    TopicAlreadyExistsLegacy(String),
    ServiceAlreadyExistsLegacy(String),
    ServiceCallTimeoutLegacy(String),
    ServiceCallFailedLegacy(String),
    ParameterLegacy(String),
    NodeLegacy(String),
    PublisherLegacy(String),
    SubscriberLegacy(String),
    ServiceLegacy(String),
    ClientLegacy(String),
    NotSupportedLegacy(String),
    NotImplementedLegacy(String),
    OtherLegacy(String),

}

Error type for Zenobuf operations

Variants

Transport

Error from the Zenoh transport layer

Fields

source: Error

context: String

MessageSerialization

Error during message serialization

Fields

source: EncodeError

type_name: &'static str

MessageDeserialization

Error during message deserialization

Fields

source: DecodeError

type_name: &'static str

Serialization(String)

Error during serialization or deserialization (legacy)

Encoding(EncodeError)

Error during message encoding (legacy)

Decoding(DecodeError)

Error during message decoding (legacy)

NodeAlreadyExists

Error when a node with the same name already exists

Fields

name: String

TopicAlreadyExists

Error when a topic with the same name already exists

Fields

topic: String

node: String

ServiceAlreadyExists

Error when a service with the same name already exists

Fields

service: String

node: String

ServiceCallTimeout

Error when a service call times out

Fields

service: String

timeout_ms: u64

ServiceCallFailed

Error when a service call fails

Fields

service: String

reason: String

Parameter

Error when a parameter operation fails

Fields

name: String

reason: String

Node

Error when a node operation fails

Fields

node: String

reason: String

Publisher

Error when a publisher operation fails

Fields

topic: String

reason: String

Subscriber

Error when a subscriber operation fails

Fields

topic: String

reason: String

Service

Error when a service operation fails

Fields

service: String

reason: String

Client

Error when a client operation fails

Fields

service: String

reason: String

NotSupported

Error when an operation is not supported

Fields

operation: String

reason: String

NotImplemented

Error when an operation is not implemented

Fields

operation: String

reason: String

Configuration

Configuration error

Fields

reason: String

Network

Network error

Fields

reason: String

Other

Other errors

Fields

reason: String

NodeAlreadyExistsLegacy(String)

TopicAlreadyExistsLegacy(String)

ServiceAlreadyExistsLegacy(String)

ServiceCallTimeoutLegacy(String)

ServiceCallFailedLegacy(String)

ParameterLegacy(String)

NodeLegacy(String)

PublisherLegacy(String)

SubscriberLegacy(String)

ServiceLegacy(String)

ClientLegacy(String)

NotSupportedLegacy(String)

NotImplementedLegacy(String)

OtherLegacy(String)

Implementations

impl Error

pub fn transport(source: Error, context: impl Into<String>) -> Self

Create a transport error with context

pub fn message_serialization( source: EncodeError, type_name: &'static str, ) -> Self

Create a message serialization error

pub fn message_deserialization( source: DecodeError, type_name: &'static str, ) -> Self

Create a message deserialization error

pub fn node_already_exists(name: impl Into<String>) -> Self

Create a node already exists error

pub fn topic_already_exists( topic: impl Into<String>, node: impl Into<String>, ) -> Self

Create a topic already exists error

pub fn service_already_exists( service: impl Into<String>, node: impl Into<String>, ) -> Self

Create a service already exists error

pub fn service_call_timeout(service: impl Into<String>, timeout_ms: u64) -> Self

Create a service call timeout error

pub fn service_call_failed( service: impl Into<String>, reason: impl Into<String>, ) -> Self

Create a service call failed error

pub fn parameter(name: impl Into<String>, reason: impl Into<String>) -> Self

Create a parameter error

pub fn node(node: impl Into<String>, reason: impl Into<String>) -> Self

Create a node error

pub fn publisher(topic: impl Into<String>, reason: impl Into<String>) -> Self

Create a publisher error

pub fn subscriber(topic: impl Into<String>, reason: impl Into<String>) -> Self

Create a subscriber error

pub fn service(service: impl Into<String>, reason: impl Into<String>) -> Self

Create a service error

pub fn client(service: impl Into<String>, reason: impl Into<String>) -> Self

Create a client error

pub fn configuration(reason: impl Into<String>) -> Self

Create a configuration error

pub fn network(reason: impl Into<String>) -> Self

Create a network error

pub fn other(reason: impl Into<String>) -> Self

Create an other error