zenobuf_core::error

Enum Error

pub enum Error {

    Zenoh(Error),
    Serialization(String),
    Encoding(EncodeError),
    Decoding(DecodeError),
    NodeAlreadyExists(String),
    TopicAlreadyExists(String),
    ServiceAlreadyExists(String),
    ServiceCallTimeout(String),
    ServiceCallFailed(String),
    Parameter(String),
    Node(String),
    Publisher(String),
    Subscriber(String),
    Service(String),
    Client(String),
    NotSupported(String),
    NotImplemented(String),
    Other(String),

}

Error type for Zenobuf operations

Variants

Zenoh(Error)

Error from the Zenoh transport layer

Serialization(String)

Error during serialization or deserialization

Encoding(EncodeError)

Error during message encoding

Decoding(DecodeError)

Error during message decoding

NodeAlreadyExists(String)

Error when a node with the same name already exists

TopicAlreadyExists(String)

Error when a topic with the same name already exists

ServiceAlreadyExists(String)

Error when a service with the same name already exists

ServiceCallTimeout(String)

Error when a service call times out

ServiceCallFailed(String)

Error when a service call fails

Parameter(String)

Error when a parameter operation fails

Node(String)

Error when a node operation fails

Publisher(String)

Error when a publisher operation fails

Subscriber(String)

Error when a subscriber operation fails

Service(String)

Error when a service operation fails

Client(String)

Error when a client operation fails

NotSupported(String)

Error when an operation is not supported

NotImplemented(String)

Error when an operation is not implemented

Other(String)

Other errors